// src/types/weatherResponse.ts
/**
 * generated by QuickType
 * https://app.quicktype.io/
 */
export interface WeatherResponse {
  success: string;
  result:  Result;
  records: Records;
}

export interface Records {
  locations: RecordsLocation[];
}

export interface RecordsLocation {
  datasetDescription: string;
  locationsName:      string;
  dataid:             string;
  location:           LocationLocation[];
}

export interface LocationLocation {
  locationName:   string;
  geocode:        string;
  lat:            string;
  lon:            string;
  weatherElement: WeatherElement[];
}

export interface WeatherElement {
  elementName: string;
  description: string;
  time:        Time[];
}

export interface Time {
  startTime:    string;
  endTime:      string;
  elementValue: ElementValue[];
}

export interface ElementValue {
  value:    string;
  measures: Measures;
}

export enum Measures {
  MeasuresNA = "NA ",
  Na = "NA",
  The8方位 = "8方位",
  公尺秒 = "公尺/秒",
  攝氏度 = "攝氏度",
  曝曬級數 = "曝曬級數",
  百分比 = "百分比",
  紫外線指數 = "紫外線指數",
  自定義Ci文字 = "自定義 CI 文字",
  自定義Wx單位 = "自定義 Wx 單位",
  自定義Wx文字 = "自定義 Wx 文字",
  蒲福風級 = "蒲福風級",
}

export interface Result {
  resource_id: string;
  fields:      Field[];
}

export interface Field {
  id:   string;
  type: Type;
}

export enum Type {
  Double = "Double",
  String = "String",
  Timestamp = "Timestamp",
}